feat(schedule): «كل كام ساعة» و«مرة واحدة»، والتذكير الأساسي قبل الإعادات لما ميزانية iOS تقصر
- «بياخده إزاي؟»: كل يوم / كل كام ساعة (٢/٣/٤/٦/٨/١٢ بيتفرد لساعات ثابتة يومية، بمعاينة
  قبل الحفظ) / مرة واحدة (DoseRepeat.once)؛ و«خليه كل كام ساعة» في تعديل الدوا
- «اليوم فقط» من الروشتة بتتحفظ «مرة واحدة»
- splitPendingBudget: لحد ٢٤ تذكير الخطة هي هي بالحرف؛ فوقه الأساسي بياخد خانات الإعادات
  لحد ٤٤ (الأقرب الأول)، والسلّم ١٤ زي ما هو — ٣ أدوية كل ٤ ساعات: ٣٢ → ٥٩ ساعة تغطية
- lowCoverage للأدمن لما الخطة تتقص والتغطية < ٤٨ ساعة (horizon_until الموجود — من غير هجرة)
- averageDosesPerDay واحدة للمخزون عند المريض والدائرة
- اختبارات قديمة اتعدّلت عن قصد: عدّاد «٢٤» في ٧ تأكيدات على مرضى ميزانيتهم قاصرة بقى ٤٤،
  واختبار إعادة شاشة القفل بقى بدوا واحد عشان الإعادات تتجدول

+++ lib/data/care/caregiver_remote.dart
@@ -65,7 +65,7 @@ class CaregiverMedication {
   final int? stockWarnDays;
 
   /// الجرعات اليومية الشغّالة — منها «فاضله كام يوم» (مش ساعات محسوبة).
-  final int dosesPerDay;
+  final double dosesPerDay;
 
   /// ٠٠٣١: «لسه ماتشترتش» — null = اتشرى. قراية بس.
   final DateTime? notBoughtAt;
